bugfix [to commit #3cb4800] : make sure the directory gets created
The commit #3cb4800 made several optimizing changes in one go,
including removal of Boost-1.64-specific scripts and substituting them
with a reference to the earlier scripts for Boost-1.62 .
However, the creation of the 'install' directory was missing
from 1.62, which had a knock-on effect on all later versions.

I'm restoring this command. Hopefully it will also fix the
already-closed issue #8 in ck-math.